requirements.txt
cloudfoundry-client
PyJWT
timedelta
requests